+++ lib/color_chart.dart
@@ -1,1 +1,2 @@
 export 'src/basic.dart';
+export 'src/utils.dart';