+++ lib/src/basic.dart
@@ -1,7 +1,7 @@
 import 'dart:ui';
 
+/// Monochrome colors from dark to light.
 extension MonoColors on Color {
-  static const transparent = Color(0x00000000);
   static const white = Color(0xffffffff);
   static const black = Color(0xff000000);
 
@@ -11,4 +11,12 @@ extension MonoColors on Color {
   static const ink = Color(0xff343637);
   static const night = Color(0xff232526);
   static const coal = Color(0xff1a1918);
+}
+
+/// Transparent monochrome colors with opacity.
+extension MaskColors on Color {
+  /// Completely transparent:
+  /// Usually used as placeholder to ensure collect event.
+  /// Otherwise, it's not recommended to use such completely transparent color.
+  static const transparent = Color(0x00000000);
 }