+++ CHANGELOG.md
@@ -1,8 +1,14 @@
+## 0.2.0
+
+- Move `transparent` into new namespace `MaskColors` (API change).
+- Hex color format convert as utils.
+
 ## 0.1.1
 
 - Migrate to [TreeInfra](https://github.com/treeinfra) org.
 - Update ci to avoid discontinued imports.
 - Use `lintall` to replace discontinued `strict_lint` package.
+- Update Flutter version code to `3.22.0`.
 
 ## 0.1.0